requirements.txt
click==6.7
Flask==0.12.2
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
paho-mqtt==1.3.0
PyMySQL==0.7.11
Werkzeug==0.12.2
